feat: moved remainig actions from app to actions dir
